Remove redundant SwtOperation
