+++ CMakeLists.txt
@@ -75,10 +75,10 @@ file(GLOB FRED
 # add_executable(AlfResponseParser-Test test/AlfResponseParser/ParsingTest.cpp mapi/src/AlfResponseParser.cpp)
 # add_executable(SwtSequence-Test test/SwtSequence/SwtSequenceTest.cpp mapi/src/SwtSequence.cpp)
 # add_executable(ParameterInterpretation-Test test/Parameter/ParameterInterpretationTest.cpp mapi/src/Parameter.cpp)
-add_executable(Parameters-Test test/Parameters/ParametersTest.cpp mapi/src/Parameters.cpp mapi/src/Parameter.cpp mapi/src/SwtOperation.cpp mapi/src/AlfResponseParser.cpp mapi/src/WinCCRequest.cpp ${FRED} core/src/Parser/utility.cpp)
+add_executable(Parameters-Test test/Parameters/ParametersTest.cpp mapi/src/Parameters.cpp mapi/src/Parameter.cpp mapi/src/AlfResponseParser.cpp mapi/src/WinCCRequest.cpp mapi/src/SwtSequence.cpp ${FRED} core/src/Parser/utility.cpp)
 # target_link_libraries(SwtSequence-Test gtest gtest_main)
 # target_link_libraries(ParameterInterpretation-Test gtest gtest_main)
-target_link_libraries(Parameters-Test Alfred clntsh occi clntshcore nnz19 ociicus gtest gtest_main)
+target_link_libraries(Parameters-Test Alfred clntsh occi nnz19 ociicus gtest gtest_main)
 
 # add_test(NAME Unit-Test-AlfResponseParser COMMAND AlfResponseParser-Test)
 # add_test(NAME Unit-Test-SwtSequence COMMAND SwtSequence-Test)
