+++ mapi/include/SwtSequence.h
@@ -17,6 +17,7 @@ class SwtSequence
 
     struct SwtOperation
     {
+        SwtOperation() = default;
         SwtOperation( Operation type, uint32_t address, std::array<uint32_t,2> data=std::array<uint32_t,2>(), bool expectResponse=false);
         Operation type;
         uint32_t address;
@@ -24,7 +25,7 @@ class SwtSequence
         bool expectResponse;
     };
 
-    
+
     SwtSequence():m_buffer("reset\n"){}
     SwtSequence(const std::vector<SwtOperation>& operations);
